+++ packages/e2b/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @computesdk/e2b
 
+## 1.7.5
+
+### Patch Changes
+
+- Updated dependencies [251f324]
+  - computesdk@1.9.6
+
+## 1.7.4
+
+### Patch Changes
+
+- Updated dependencies [b027cd9]
+  - computesdk@1.9.5
+
+## 1.7.3
+
+### Patch Changes
+
+- computesdk@1.9.4
+
 ## 1.7.2
 
 ### Patch Changes